+++ session_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/lucas-clemente/quic-go/internal/crypto"
 	"github.com/lucas-clemente/quic-go/internal/handshake"
 	"github.com/lucas-clemente/quic-go/internal/mocks"
+	"github.com/lucas-clemente/quic-go/internal/mocks/ackhandler"
 	"github.com/lucas-clemente/quic-go/internal/protocol"
 	"github.com/lucas-clemente/quic-go/internal/testdata"
 	"github.com/lucas-clemente/quic-go/internal/wire"
@@ -277,7 +278,7 @@ var _ = Describe("Session", func() {
 		Context("handling ACK frames", func() {
 			It("informs the SentPacketHandler about ACKs", func() {
 				f := &wire.AckFrame{LargestAcked: 3, LowestAcked: 2}
-				sph := mocks.NewMockSentPacketHandler(mockCtrl)
+				sph := mockackhandler.NewMockSentPacketHandler(mockCtrl)
 				sph.EXPECT().ReceivedAck(f, protocol.PacketNumber(42), protocol.EncryptionSecure, gomock.Any())
 				sph.EXPECT().GetLowestPacketNotConfirmedAcked()
 				sess.sentPacketHandler = sph
@@ -287,11 +288,11 @@ var _ = Describe("Session", func() {
 			})
 
 			It("tells the ReceivedPacketHandler to ignore low ranges", func() {
-				sph := mocks.NewMockSentPacketHandler(mockCtrl)
+				sph := mockackhandler.NewMockSentPacketHandler(mockCtrl)
 				sph.EXPECT().ReceivedAck(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
 				sph.EXPECT().GetLowestPacketNotConfirmedAcked().Return(protocol.PacketNumber(0x42))
 				sess.sentPacketHandler = sph
-				rph := mocks.NewMockReceivedPacketHandler(mockCtrl)
+				rph := mockackhandler.NewMockReceivedPacketHandler(mockCtrl)
 				rph.EXPECT().IgnoreBelow(protocol.PacketNumber(0x42))
 				sess.receivedPacketHandler = rph
 				err := sess.handleAckFrame(&wire.AckFrame{LargestAcked: 3, LowestAcked: 2}, protocol.EncryptionUnencrypted)
@@ -800,59 +801,22 @@ var _ = Describe("Session", func() {
 			packetNumber := protocol.PacketNumber(0x035e)
 			err := sess.receivedPacketHandler.ReceivedPacket(packetNumber, true)
 			Expect(err).ToNot(HaveOccurred())
-			err = sess.sendPacket()
+			packetSent, err := sess.sendPacket()
 			Expect(err).NotTo(HaveOccurred())
+			Expect(packetSent).To(BeTrue())
 			Expect(mconn.written).To(HaveLen(1))
 			Expect(mconn.written).To(Receive(ContainSubstring(string([]byte{0x03, 0x5e}))))
 		})
 
-		It("sends ACK frames when congestion limited", func() {
-			swf := &wire.StopWaitingFrame{LeastUnacked: 10}
-			sph := mocks.NewMockSentPacketHandler(mockCtrl)
-			sph.EXPECT().GetLeastUnacked().AnyTimes()
-			sph.EXPECT().SendingAllowed().Return(false)
-			sph.EXPECT().GetStopWaitingFrame(false).Return(swf)
-			sph.EXPECT().SentPacket(gomock.Any()).Do(func(p *ackhandler.Packet) {
-				Expect(p.Frames).To(HaveLen(2))
-				Expect(p.Frames[0]).To(BeAssignableToTypeOf(&wire.AckFrame{}))
-				Expect(p.Frames[1]).To(Equal(swf))
-			})
-			sess.sentPacketHandler = sph
-			sess.packer.packetNumberGenerator.next = 0x1338
-			sess.receivedPacketHandler.ReceivedPacket(1, true)
-			err := sess.sendPacket()
-			Expect(err).NotTo(HaveOccurred())
-			Expect(mconn.written).To(HaveLen(1))
-		})
-
-		It("doesn't include a STOP_WAITING for an ACK-only packet for IETF QUIC", func() {
-			sess.version = versionIETFFrames
-			sph := mocks.NewMockSentPacketHandler(mockCtrl)
-			sph.EXPECT().GetLeastUnacked().AnyTimes()
-			sph.EXPECT().SendingAllowed().Return(false)
-			sph.EXPECT().SentPacket(gomock.Any()).Do(func(p *ackhandler.Packet) {
-				Expect(p.Frames).To(HaveLen(1))
-				Expect(p.Frames[0]).To(BeAssignableToTypeOf(&wire.AckFrame{}))
-			})
-			sess.sentPacketHandler = sph
-			sess.packer.packetNumberGenerator.next = 0x1338
-			sess.receivedPacketHandler.ReceivedPacket(1, true)
-			err := sess.sendPacket()
-			Expect(err).NotTo(HaveOccurred())
-			Expect(mconn.written).To(HaveLen(1))
-		})
-
 		It("sends a retransmittable packet when required by the SentPacketHandler", func() {
 			sess.packer.QueueControlFrame(&wire.AckFrame{LargestAcked: 1000})
-			sph := mocks.NewMockSentPacketHandler(mockCtrl)
+			sph := mockackhandler.NewMockSentPacketHandler(mockCtrl)
 			sph.EXPECT().GetLeastUnacked().AnyTimes()
-			sph.EXPECT().SendingAllowed().Return(true)
-			sph.EXPECT().SendingAllowed().Return(false)
 			sph.EXPECT().DequeuePacketForRetransmission()
 			sph.EXPECT().ShouldSendRetransmittablePacket().Return(true)
 			sph.EXPECT().SentPacket(gomock.Any())
 			sess.sentPacketHandler = sph
-			err := sess.sendPacket()
+			_, err := sess.sendPacket()
 			Expect(err).ToNot(HaveOccurred())
 			Expect(mconn.written).To(HaveLen(1))
 		})
@@ -862,10 +826,8 @@ var _ = Describe("Session", func() {
 			fc.EXPECT().GetWindowUpdate().Return(protocol.ByteCount(0x1337))
 			fc.EXPECT().IsNewlyBlocked()
 			sess.connFlowController = fc
-			sph := mocks.NewMockSentPacketHandler(mockCtrl)
+			sph := mockackhandler.NewMockSentPacketHandler(mockCtrl)
 			sph.EXPECT().GetLeastUnacked().AnyTimes()
-			sph.EXPECT().SendingAllowed().Return(true)
-			sph.EXPECT().SendingAllowed()
 			sph.EXPECT().DequeuePacketForRetransmission()
 			sph.EXPECT().ShouldSendRetransmittablePacket()
 			sph.EXPECT().SentPacket(gomock.Any()).Do(func(p *ackhandler.Packet) {
@@ -874,7 +836,7 @@ var _ = Describe("Session", func() {
 				}))
 			})
 			sess.sentPacketHandler = sph
-			err := sess.sendPacket()
+			_, err := sess.sendPacket()
 			Expect(err).ToNot(HaveOccurred())
 		})
 
@@ -883,17 +845,15 @@ var _ = Describe("Session", func() {
 				StreamID:   2,
 				ByteOffset: 20,
 			})
-			sph := mocks.NewMockSentPacketHandler(mockCtrl)
+			sph := mockackhandler.NewMockSentPacketHandler(mockCtrl)
 			sph.EXPECT().GetLeastUnacked().AnyTimes()
-			sph.EXPECT().SendingAllowed().Return(true)
-			sph.EXPECT().SendingAllowed()
 			sph.EXPECT().DequeuePacketForRetransmission()
 			sph.EXPECT().ShouldSendRetransmittablePacket()
 			sph.EXPECT().SentPacket(gomock.Any()).Do(func(p *ackhandler.Packet) {
 				Expect(p.Frames).To(ContainElement(&wire.MaxStreamDataFrame{StreamID: 2, ByteOffset: 20}))
 			})
 			sess.sentPacketHandler = sph
-			err := sess.sendPacket()
+			_, err := sess.sendPacket()
 			Expect(err).ToNot(HaveOccurred())
 		})
 
@@ -902,10 +862,8 @@ var _ = Describe("Session", func() {
 			fc.EXPECT().GetWindowUpdate()
 			fc.EXPECT().IsNewlyBlocked().Return(true, protocol.ByteCount(1337))
 			sess.connFlowController = fc
-			sph := mocks.NewMockSentPacketHandler(mockCtrl)
+			sph := mockackhandler.NewMockSentPacketHandler(mockCtrl)
 			sph.EXPECT().GetLeastUnacked().AnyTimes()
-			sph.EXPECT().SendingAllowed().Return(true)
-			sph.EXPECT().SendingAllowed()
 			sph.EXPECT().DequeuePacketForRetransmission()
 			sph.EXPECT().ShouldSendRetransmittablePacket()
 			sph.EXPECT().SentPacket(gomock.Any()).Do(func(p *ackhandler.Packet) {
@@ -914,8 +872,43 @@ var _ = Describe("Session", func() {
 				}))
 			})
 			sess.sentPacketHandler = sph
-			err := sess.sendPacket()
+			_, err := sess.sendPacket()
 			Expect(err).ToNot(HaveOccurred())
+		})
+
+		FIt("paces multiple packets", func() {
+			sess.queueControlFrame(&wire.MaxDataFrame{ByteOffset: 1})
+			sph := mockackhandler.NewMockSentPacketHandler(mockCtrl)
+			sph.EXPECT().DequeuePacketForRetransmission().Times(2)
+			sph.EXPECT().GetAlarmTimeout().AnyTimes()
+			sph.EXPECT().GetLeastUnacked().AnyTimes()
+			sph.EXPECT().ShouldSendRetransmittablePacket().Times(2)
+			sph.EXPECT().SentPacket(gomock.Any()).Times(2)
+			sph.EXPECT().TimeUntilSend(gomock.Any()).Do(func(now time.Time) {
+				Expect(now).To(BeTemporally("~", time.Now(), time.Second))
+			}).Return(1, 200*time.Millisecond) // is called every time a packet is sent
+			sph.EXPECT().TimeUntilSend(gomock.Any()).Do(func(now time.Time) {
+				Expect(now).To(BeTemporally("~", time.Now(), time.Second))
+			}).Return(1, time.Hour)
+			sph.EXPECT().SendingAllowed().Do(func() { // is called before and after sending a packet
+				// make sure there's something to send
+				sess.queueControlFrame(&wire.MaxDataFrame{ByteOffset: 2})
+			}).Return(true).Times(4) // allow 2 packets...
+			sph.EXPECT().SendingAllowed() // ...then report that we're congestion limited
+			sess.sentPacketHandler = sph
+			done := make(chan struct{})
+			go func() {
+				defer GinkgoRecover()
+				sess.run()
+				close(done)
+			}()
+			// sess.scheduleSending()
+			Eventually(mconn.written).Should(HaveLen(1))
+			Consistently(mconn.written, 100*time.Millisecond).Should(HaveLen(1))
+			Eventually(mconn.written, 200*time.Millisecond).Should(HaveLen(2))
+			// make the go routine return
+			sess.Close(nil)
+			Eventually(done).Should(BeClosed())
 		})
 
 		It("sends public reset", func() {
@@ -931,11 +924,9 @@ var _ = Describe("Session", func() {
 				Data:     []byte("foobar"),
 			}
 			var sentPacket *ackhandler.Packet
-			sph := mocks.NewMockSentPacketHandler(mockCtrl)
+			sph := mockackhandler.NewMockSentPacketHandler(mockCtrl)
 			sph.EXPECT().GetLeastUnacked().AnyTimes()
 			sph.EXPECT().GetStopWaitingFrame(gomock.Any())
-			sph.EXPECT().SendingAllowed().Return(true)
-			sph.EXPECT().SendingAllowed().Return(false)
 			sph.EXPECT().DequeuePacketForRetransmission()
 			sph.EXPECT().ShouldSendRetransmittablePacket()
 			sph.EXPECT().SentPacket(gomock.Any()).Do(func(p *ackhandler.Packet) {
@@ -948,7 +939,7 @@ var _ = Describe("Session", func() {
 			sess.streamFramer.AddFrameForRetransmission(f)
 			_, err := sess.GetOrOpenStream(5)
 			Expect(err).ToNot(HaveOccurred())
-			err = sess.sendPacket()
+			_, err = sess.sendPacket()
 			Expect(err).NotTo(HaveOccurred())
 			Expect(mconn.written).To(HaveLen(1))
 			Expect(sentPacket.PacketNumber).To(Equal(protocol.PacketNumber(0x1337 + 9)))
@@ -958,16 +949,78 @@ var _ = Describe("Session", func() {
 		})
 	})
 
+	Context("sending ACK only packets", func() {
+		It("doesn't do anything if there's no ACK to be sent", func() {
+			sph := mockackhandler.NewMockSentPacketHandler(mockCtrl)
+			sess.sentPacketHandler = sph
+			err := sess.maybeSendAckOnlyPacket()
+			Expect(err).ToNot(HaveOccurred())
+			Expect(mconn.written).To(BeEmpty())
+		})
+
+		It("sends ACK only packets", func() {
+			swf := &wire.StopWaitingFrame{LeastUnacked: 10}
+			sph := mockackhandler.NewMockSentPacketHandler(mockCtrl)
+			sph.EXPECT().GetLeastUnacked()
+			sph.EXPECT().GetAlarmTimeout().AnyTimes()
+			sph.EXPECT().SendingAllowed()
+			sph.EXPECT().GetStopWaitingFrame(false).Return(swf)
+			sph.EXPECT().SentPacket(gomock.Any()).Do(func(p *ackhandler.Packet) {
+				Expect(p.Frames).To(HaveLen(2))
+				Expect(p.Frames[0]).To(BeAssignableToTypeOf(&wire.AckFrame{}))
+				Expect(p.Frames[1]).To(Equal(swf))
+			})
+			sess.sentPacketHandler = sph
+			sess.packer.packetNumberGenerator.next = 0x1338
+			sess.receivedPacketHandler.ReceivedPacket(1, true)
+			done := make(chan struct{})
+			go func() {
+				defer GinkgoRecover()
+				sess.run()
+				close(done)
+			}()
+			sess.scheduleSending()
+			Eventually(mconn.written).Should(HaveLen(1))
+			// make sure that the go routine returns
+			sess.Close(nil)
+			Eventually(done).Should(BeClosed())
+		})
+
+		It("doesn't include a STOP_WAITING for an ACK-only packet for IETF QUIC", func() {
+			sess.version = versionIETFFrames
+			done := make(chan struct{})
+			sph := mockackhandler.NewMockSentPacketHandler(mockCtrl)
+			sph.EXPECT().GetLeastUnacked()
+			sph.EXPECT().GetAlarmTimeout().AnyTimes()
+			sph.EXPECT().SendingAllowed()
+			sph.EXPECT().SentPacket(gomock.Any()).Do(func(p *ackhandler.Packet) {
+				Expect(p.Frames).To(HaveLen(1))
+				Expect(p.Frames[0]).To(BeAssignableToTypeOf(&wire.AckFrame{}))
+			})
+			sess.sentPacketHandler = sph
+			sess.packer.packetNumberGenerator.next = 0x1338
+			sess.receivedPacketHandler.ReceivedPacket(1, true)
+			go func() {
+				defer GinkgoRecover()
+				sess.run()
+				close(done)
+			}()
+			sess.scheduleSending()
+			Eventually(mconn.written).Should(HaveLen(1))
+			// make sure that the go routine returns
+			sess.Close(nil)
+			Eventually(done).Should(BeClosed())
+		})
+	})
+
 	Context("retransmissions", func() {
-		var sph *mocks.MockSentPacketHandler
+		var sph *mockackhandler.MockSentPacketHandler
 		BeforeEach(func() {
 			// a STOP_WAITING frame is added, so make sure the packet number of the new package is higher than the packet number of the retransmitted packet
 			sess.packer.packetNumberGenerator.next = 0x1337 + 10
 			sess.packer.hasSentPacket = true // make sure this is not the first packet the packer sends
-			sph = mocks.NewMockSentPacketHandler(mockCtrl)
+			sph = mockackhandler.NewMockSentPacketHandler(mockCtrl)
 			sph.EXPECT().GetLeastUnacked().AnyTimes()
-			sph.EXPECT().SendingAllowed().Return(true)
-			sph.EXPECT().ShouldSendRetransmittablePacket()
 			sess.sentPacketHandler = sph
 			sess.packer.cryptoSetup = &mockCryptoSetup{encLevelSeal: protocol.EncryptionForwardSecure}
 		})
@@ -981,15 +1034,14 @@ var _ = Describe("Session", func() {
 					Frames:          []wire.Frame{sf},
 					EncryptionLevel: protocol.EncryptionUnencrypted,
 				})
-				sph.EXPECT().DequeuePacketForRetransmission()
 				sph.EXPECT().SentPacket(gomock.Any()).Do(func(p *ackhandler.Packet) {
 					Expect(p.EncryptionLevel).To(Equal(protocol.EncryptionUnencrypted))
 					Expect(p.Frames).To(Equal([]wire.Frame{swf, sf}))
 				})
-				err := sess.sendPacket()
+				sentPacket, err := sess.sendPacket()
 				Expect(err).ToNot(HaveOccurred())
+				Expect(sentPacket).To(BeTrue())
 				Expect(mconn.written).To(HaveLen(1))
-
 			})
 
 			It("retransmits an unencrypted packet, and doesn't add a STOP_WAITING frame (for IETF QUIC)", func() {
@@ -1000,13 +1052,13 @@ var _ = Describe("Session", func() {
 					Frames:          []wire.Frame{sf},
 					EncryptionLevel: protocol.EncryptionUnencrypted,
 				})
-				sph.EXPECT().DequeuePacketForRetransmission()
 				sph.EXPECT().SentPacket(gomock.Any()).Do(func(p *ackhandler.Packet) {
 					Expect(p.EncryptionLevel).To(Equal(protocol.EncryptionUnencrypted))
 					Expect(p.Frames).To(Equal([]wire.Frame{sf}))
 				})
-				err := sess.sendPacket()
+				packetSent, err := sess.sendPacket()
 				Expect(err).ToNot(HaveOccurred())
+				Expect(packetSent).To(BeTrue())
 				Expect(mconn.written).To(HaveLen(1))
 			})
 
@@ -1018,8 +1070,10 @@ var _ = Describe("Session", func() {
 						Frames:          []wire.Frame{sf},
 						EncryptionLevel: protocol.EncryptionSecure,
 					})
+				sph.EXPECT().ShouldSendRetransmittablePacket()
 				sph.EXPECT().DequeuePacketForRetransmission()
-				err := sess.sendPacket()
+				packetSent, err := sess.sendPacket()
+				Expect(packetSent).To(BeFalse())
 				Expect(err).ToNot(HaveOccurred())
 				Expect(mconn.written).To(BeEmpty())
 			})
@@ -1043,8 +1097,9 @@ var _ = Describe("Session", func() {
 					Expect(p.Frames).To(Equal([]wire.Frame{swf, f}))
 					Expect(p.EncryptionLevel).To(Equal(protocol.EncryptionForwardSecure))
 				})
-				sph.EXPECT().SendingAllowed()
-				err := sess.sendPacket()
+				sph.EXPECT().ShouldSendRetransmittablePacket()
+				packetSent, err := sess.sendPacket()
+				Expect(packetSent).To(BeTrue())
 				Expect(err).NotTo(HaveOccurred())
 				Expect(mconn.written).To(HaveLen(1))
 			})
@@ -1065,13 +1120,14 @@ var _ = Describe("Session", func() {
 					Expect(p.Frames).To(Equal([]wire.Frame{f}))
 					Expect(p.EncryptionLevel).To(Equal(protocol.EncryptionForwardSecure))
 				})
-				sph.EXPECT().SendingAllowed()
-				err := sess.sendPacket()
+				sph.EXPECT().ShouldSendRetransmittablePacket()
+				packetSent, err := sess.sendPacket()
 				Expect(err).NotTo(HaveOccurred())
+				Expect(packetSent).To(BeTrue())
 				Expect(mconn.written).To(HaveLen(1))
 			})
 
-			It("sends a StreamFrame from a packet queued for retransmission", func() {
+			It("sends a STREAM frame from a packet queued for retransmission", func() {
 				f1 := wire.StreamFrame{
 					StreamID: 0x5,
 					Data:     []byte("foobar"),
@@ -1093,12 +1149,13 @@ var _ = Describe("Session", func() {
 				sph.EXPECT().DequeuePacketForRetransmission().Return(p1)
 				sph.EXPECT().DequeuePacketForRetransmission().Return(p2)
 				sph.EXPECT().DequeuePacketForRetransmission()
+				sph.EXPECT().ShouldSendRetransmittablePacket()
 				sph.EXPECT().GetStopWaitingFrame(true).Return(&wire.StopWaitingFrame{})
 				sph.EXPECT().SentPacket(gomock.Any())
-				sph.EXPECT().SendingAllowed()
 
-				err := sess.sendPacket()
+				packetSent, err := sess.sendPacket()
 				Expect(err).NotTo(HaveOccurred())
+				Expect(packetSent).To(BeTrue())
 				Expect(mconn.written).To(HaveLen(1))
 				packet := <-mconn.written
 				Expect(packet).To(ContainSubstring("foobar"))
@@ -1158,7 +1215,7 @@ var _ = Describe("Session", func() {
 		})
 
 		It("sets the timer to the ack timer", func() {
-			rph := mocks.NewMockReceivedPacketHandler(mockCtrl)
+			rph := mockackhandler.NewMockReceivedPacketHandler(mockCtrl)
 			rph.EXPECT().GetAckFrame().Return(&wire.AckFrame{LargestAcked: 0x1337})
 			rph.EXPECT().GetAlarmTimeout().Return(time.Now().Add(10 * time.Millisecond)).MinTimes(1)
 			sess.receivedPacketHandler = rph